pull out highlight and attach
