+++ lib/taiko.js
@@ -22,8 +22,7 @@ const browserHandler = require('./handlers/browserHandler');
 const emulationHandler = require('./handlers/emulationHandler');
 const { description } = require('./actions/pageActionChecks');
 const { match, $$xpath, findElements, findFirstElement } = require('./elementSearch');
-const { handleRelativeSearch, RelativeSearchElement } = require('./proximityElementSearch');
-const Element = require('./elements/element');
+const { RelativeSearchElement } = require('./proximityElementSearch');
 const { scrollToElement } = require('./actions/scrollTo');
 const {
   setConfig,
@@ -1089,21 +1088,9 @@ module.exports.clear = async (selector, options = {}) => {
  */
 module.exports.attach = async (filepath, to) => {
   validate();
-  let resolvedPath = filepath ? path.resolve(process.cwd(), filepath) : path.resolve(process.cwd());
-  if (!fs.existsSync(resolvedPath)) {
-    throw new Error(`File ${resolvedPath} does not exist.`);
-  }
-  if (isString(to)) {
-    to = fileField(to);
-  } else if (!isSelector(to) && !isElement(to)) {
-    throw Error('Invalid element passed as parameter');
-  }
-  const element = await findFirstElement(to);
-  if (defaultConfig.headful) {
-    await highlightElement(element);
-  }
-  await domHandler.setFileInputFiles(element.get(), resolvedPath);
-  descEvent.emit('success', 'Attached ' + resolvedPath + ' to the ' + description(to, true));
+  const { attach } = require('./actions/attach');
+  const desc = await attach(filepath, to);
+  descEvent.emit('success', desc);
 };
 
 /**
@@ -1174,35 +1161,9 @@ async function highlight(selector, ...args) {
     console.warn('Highlights are disabled. Please enable highlights.');
     return;
   }
-
-  function highlightNode() {
-    function addTaikoHighlightStyleIfNotPresent(self) {
-      let taikoHighlightStyleID = 'taiko_highlight';
-      const root = self.getRootNode();
-      if (root.getElementById(taikoHighlightStyleID)) {
-        return;
-      }
-      let head =
-        root instanceof ShadowRoot
-          ? root
-          : document.head || document.getElementsByTagName('head')[0];
-      let style = document.createElement('style');
-      let css = '.taiko_highlight_style { outline: 0.5em solid red; }';
-      head.appendChild(style);
-      style.type = 'text/css';
-      style.id = taikoHighlightStyleID;
-      style.appendChild(document.createTextNode(css));
-    }
-    addTaikoHighlightStyleIfNotPresent(this);
-    let _class = 'taiko_highlight_style';
-    if (this.nodeType === Node.TEXT_NODE) {
-      return this.parentElement.classList.add(_class);
-    }
-    this.classList.add(_class);
-  }
-  let elems = await handleRelativeSearch(await findElements(selector), args);
-  await evaluate(elems[0], highlightNode);
-  descEvent.emit('success', 'Highlighted the ' + description(elems[0], true));
+  const { highlight } = require('./actions/highlight');
+  const desc = await highlight(selector, args);
+  descEvent.emit('success', desc);
 }
 
 /**
@@ -1216,20 +1177,9 @@ async function highlight(selector, ...args) {
 
 module.exports.clearHighlights = async () => {
   validate();
-  let elem = (await $$xpath('//*'))[0];
-  await evaluate(elem, function () {
-    let _class = 'taiko_highlight_style';
-    let elems = document.getElementsByClassName(_class);
-    Array.from(elems).forEach((e) => e.classList.remove(_class));
-    const searchElements = document.querySelectorAll('*');
-    for (const element of searchElements) {
-      if (element.shadowRoot) {
-        let elems = element.shadowRoot.querySelectorAll(`.${_class}`);
-        Array.from(elems).forEach((e) => e.classList.remove(_class));
-      }
-    }
-  });
-  descEvent.emit('success', 'Cleared all highlights for this page.');
+  const { clearHighlights } = require('./actions/highlight');
+  const desc = await clearHighlights();
+  descEvent.emit('success', desc);
 };
 
 /**
@@ -2652,12 +2602,6 @@ const dialog = (dialogType, dialogMessage, callback) => {
 const createJsDialogEvent = (message, dType) => {
   let hash = crypto.createHash('md5').update(message.toString()).digest('hex');
   return dType + '_' + hash;
-};
-
-const evaluate = async (selector, func) => {
-  let objectId = (await findFirstElement(selector)).get();
-  let { result } = await runtimeHandler.runtimeCallFunctionOn(func, null, { objectId: objectId });
-  return result || {};
 };
 
 const rectangle = async (selector, callback) => {
